+++ Misc/NEWS.d/next/IDLE/2019-03-21-08-35-00.bpo-36390.OdDCGk.rst
@@ -1,2 +1,2 @@
-Rename paragraph.py to format.py and add region formatting methods
-from editor.py.  Add tests for the latter.
+Gather Format menu functions into format.py.  Combine
+paragraph.py, rstrip.py, and format methods from editor.py.
